Добавлены констант и утилз, favorite, null вместо кавычек, комменты в объекте карточки
