+++ src/components/rating.js
@@ -1,22 +1,3 @@
-import {RATING} from '../const.js';
-
-/**
- * Получает звание пользователя
- * @param  {number} count {кол-во просмотренных фильмов}
- * @return {string} {звание пользователя}
- */
-const getRating = (count) => {
-  let result = ``;
-  if (count > 1 && count <= 10) {
-    result = RATING[0];
-  } else if (count > 10 && count <= 20) {
-    result = RATING[1];
-  } else if (count > 21) {
-    result = RATING[2];
-  }
-  return result;
-};
-
 /**
  * Создает html-код блока звания пользователя по шаблону
  * @param  {string} rating {звание пользователя}
@@ -31,4 +12,4 @@ const createRatingTemplate = (rating) => {
   );
 };
 
-export {getRating, createRatingTemplate};
+export {createRatingTemplate};
