+++ src/mock/card.js
@@ -1,66 +1,13 @@
-import {CARDS_COUNT} from '../const.js';
+import {CARDS_COUNT, POSTER_ITEMS, GENRE_ITEMS, DESCRIPTION_ITEMS, TITLE_ITEMS} from '../const.js';
 import {getRandomArrayItem, getRandomBoolean, getRandomDecimalNumber, getRandomIntegerNumber} from '../utils.js';
-
-const TitleItems = [
-  `101 долматинец`,
-  `Вики, Кристина, Барселона`,
-  `Поймай меня, если сможешь`,
-  `Тупой и ещё тупее`,
-  `Крестный отец`,
-  `Один дома`,
-  `Тот самый Мюнхгаузен`,
-  `Всегда говори да`,
-  `Бойцовский клуб`,
-  `Красавица и чудовище`,
-  `Хороший, плохой, злой`,
-  `Крепкий орешек`,
-  `Любовь и голуби`,
-  `Безумно влюблённый`,
-  `Старик Хоттабыч`,
-];
-
-const PosterItems = [
-  `./images/posters/made-for-each-other.png`,
-  `./images/posters/popeye-meets-sinbad.png`,
-  `./images/posters/sagebrush-trail.jpg`,
-  `./images/posters/santa-claus-conquers-the-martians.jpg`,
-  `./images/posters/the-dance-of-life.jpg`,
-  `./images/posters/the-great-flamarion.jpg`,
-  `./images/posters/the-man-with-the-golden-arm.jpg`,
-];
-
-const DescriptionItems = [
-  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
-  `Fusce tristique felis at fermentum pharetra.`,
-  `Aliquam id orci ut lectus varius viverra.`,
-  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-  `Aliquam erat volutpat.`,
-  `Nunc fermentum tortor ac porta dapibus.`,
-  `In rutrum ac purus sit amet tempus.`,
-];
-
-const GenreItems = [
-  `Боевик`,
-  `Вестерн`,
-  `Драма`,
-  `Исторический`,
-  `Комедия`,
-  `Мультик`,
-  `Триллер`,
-  `Приключения`,
-  `Фантастика`,
-];
+import {getComments} from '../mock/comment.js';
 
 /**
  * Генерирует название фильма
  * @return {string} {название фильма}
  */
 const generateTitle = () => {
-  let title = getRandomArrayItem(TitleItems).split(` `);
+  let title = getRandomArrayItem(TITLE_ITEMS).split(` `);
   const randomIndex = getRandomIntegerNumber(0, title.length - 1);
   title[randomIndex] = title[randomIndex].endsWith(`,`) ? `Олег, ` : `Олег`;
   return title.join(` `);
@@ -75,7 +22,7 @@ const getDuration = () => {
   const minutes = `${getRandomIntegerNumber(1, 59)}m`;
   let duration = `${hours}${minutes}`;
   if (hours === `0h `) {
-    duration = minutes;
+    return minutes;
   }
   return duration;
 };
@@ -85,10 +32,10 @@ const getDuration = () => {
  * @return {string} описание фильма}
  */
 const getDescription = () => {
-  let description = ``;
+  let description = null;
   const descriptionLength = getRandomIntegerNumber(1, 3);
   for (let i = 0; i < descriptionLength; i++) {
-    description += ` ` + getRandomArrayItem(DescriptionItems);
+    description += ` ` + getRandomArrayItem(DESCRIPTION_ITEMS);
   }
   return description;
 };
@@ -100,7 +47,7 @@ const getDescription = () => {
 const generateCard = () => {
   const title = generateTitle();
   const year = getRandomIntegerNumber(1970, 2019);
-  const genres = GenreItems.filter(() => Math.random() > 0.5);
+  const genres = GENRE_ITEMS.filter(() => Math.random() > 0.5);
 
   return {
     title,
@@ -108,12 +55,12 @@ const generateCard = () => {
     year,
     duration: getDuration(),
     genres,
-    poster: getRandomArrayItem(PosterItems),
+    poster: getRandomArrayItem(POSTER_ITEMS),
     description: getDescription(),
-    comments: getRandomIntegerNumber(0, 20),
+    comments: getComments(getRandomIntegerNumber(0, 20)),
     isAddedToWatchlist: getRandomBoolean(),
     isWatched: getRandomBoolean(),
-    isFavourite: getRandomBoolean(),
+    isFavorite: getRandomBoolean(),
     originalTitle: title,
     director: `Anthony Mann`,
     writers: [`Anne Wigton`, `Heinz Herald`, `Richard Weil`],
@@ -130,7 +77,7 @@ const generateCard = () => {
  */
 const generateCards = () => {
   return new Array(CARDS_COUNT)
-    .fill(``)
+    .fill(null)
     .map(generateCard);
 };
